refactor(voice): extract reusable Pocket primitives
